Lots of changes in this commit.
Added horizontal and vertical word-validation
to scrabble.py, which can be used at the time
the user finalizes the hand.

Removed 'do it' button -- it was just for testing

Refactored scrabble_dfa.py to create a DFA class
that provides the ability to check if a word is
legal in the Scrabble dictionary.

Created play_game.py as a wrapper to the project.
The purpose is to ensure that Python 2.x is used
to run the project rather than 3.x, as some of
the imported modules have changed from 2.x
(current version in use) to 3.x and would not
work properly as things are currently implemented.