Modifications to word.py
This change refactors the Word class in word.py to make things
easier for the ScrabbleAI class. The Word class should now be
initialized with a word, letters_by_coord, and score.
The class is a dumb property bag that now provides overloaded
comparison operators for <, <=, !=, ==, >=, and >.
These overloads compare each Word object's score.